Commit Update Newest & feature

+++ admin/includes/view_all_comments.php
@@ -50,7 +50,7 @@
           echo "<td>{$comment_date}</td>";
           echo "<td><a href='./comments.php?approve={$comment_id}'>Approve</a></td>";
           echo "<td><a href='./comments.php?unapprove={$comment_id}'>Unapprove</a></td>";
-          echo "<td><a href='./comments.php?delete={$comment_id}'>Delete</a></td>";
+          echo "<td><a  onClick=\"javascript: return confirm('Delete confirmation') \" href='./comments.php?delete={$comment_id}'>Delete</a></td>";
           echo "</tr>";
         }
        ?>
